+++ src/source/image_source.ts
@@ -246,6 +246,8 @@ export class ImageSource extends Evented<SourceEventType> implements Source {
                     this.coordinates = newCoordinates;
                 }
                 this._finishLoading();
+            } else if (image) {
+                this.fire(new ErrorEvent(new Error(`Could not load image ${this.url}: the response is empty`)));
             }
         } catch (err) {
             // In case of abort error, the aborter may have started a new request so we don't want to clear its abort controller.